fix(config): Rename lib/config.go to lib/config/config.go and change Ollama_Port type from int16 to int

+++ lib/config/config.go
@@ -9,7 +9,7 @@ import (
 
 type Config struct {
 	Model       string
-	Ollama_Port int16
+	Ollama_Port int
 	Temperature float32
 }
